Valmistaudu lukion opetussuunnitelman 2019 lisäämiseen;
Siirrä edellisen opetussunnitelman mukaiset luokat omaan
tiedostoonsa ja jätä vanhaan tiedostoon vain opetussuunnitelmille
yhteiset osuudet. Uudelleennimeä arviointiluokat, koska
niitä tullaan käyttämään myös muun kuin kurssien kanssa.
